requirement.txt
# Everything needed in project

# Flask
# ------------------------------------------------------------------------------
click>=7.0
Flask==3.1.0
Werkzeug==3.1.3

# Database
# ------------------------------------------------------------------------------
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.39

# Migrations
# ------------------------------------------------------------------------------
Flask-Migrate==4.1.0

# Validate
# ------------------------------------------------------------------------------
email-validator==2.2.0

# Auth
# ------------------------------------------------------------------------------
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1

# Environment variable parsing
# ------------------------------------------------------------------------------
environs==14.1.1

# Debug toolbar
# ------------------------------------------------------------------------------
Flask-DebugToolbar==0.16.0

# Code quality
# ------------------------------------------------------------------------------
pre-commit==4.1.0
ruff==0.8.2

# Rest
# ------------------------------------------------------------------------------
marshmallow==3.26.1

# Swagger
# ------------------------------------------------------------------------------
flask-smorest==0.45.0

# Test
# ------------------------------------------------------------------------------
pytest==8.3.5
factory_boy==3.3.3
